requirements.txt
python-telegram-bot
google-generativeai
requests
SpeechRecognition
PyAudio
gTTS
pydub       # <-- Добавлено
simpleaudio # <-- Добавлено